+++ test/features/abstract.js
@@ -26,7 +26,14 @@ define(function () {
         $.emit('parsley:field:foo', 'bar');
         $.emit('parsley:field:foo', parsleyField, 'baz');
       });
-      it.skip('should use unsubscribe()');
+      it('should use unsubscribe()', function () {
+        $('body').append('<input type="email" id="element" />');
+        var parsleyInstance = $('#element').parsley();
+        parsleyInstance.subscribe('foo', function () { return 'bar'; });
+        expect($.subscribed().foo.length).to.be(1);
+        parsleyInstance.unsubscribe('foo');
+        expect($.subscribed().foo.length).to.be(0);
+      });
       it('should use reset() on field', function () {
         $('body').append('<input type="email" data-parsley-pattern="[A-F][0-9]{5}" data-parsley-required id="element" />');
         var parsleyField = new Parsley('#element');
@@ -77,14 +84,36 @@ define(function () {
             '<textarea id="field2"></textarea>' +
           '</form>');
         var parsleyForm = new Parsley($('#element'));
+        var fieldEventsCount = 0;
 
         // Test that a subscribed field event on parent form would be triggered by fields too
         // Here we only have field1 and field2 as valid parsley fields
         parsleyForm.subscribe('parsley:field:destroy', function () {
-          if (1 === triggered)
-            done();
+          fieldEventsCount++;
+        });
 
-          triggered++;
+        parsleyForm.subscribe('parsley:form:destroy', function () {
+          // we properly triggered before the 2 field:destroy events for this form
+          expect(fieldEventsCount).to.be(2);
+
+          // we should never enter here since parsley form instance is destroyed
+          $.listen('parsley:form:validate', function () {
+            expect(true).to.be(false);
+          });
+
+          // test that a submit event does not trigger parsley validation anymore
+          $('#element').on('submit', function (e) {
+            e.preventDefault();
+
+            expect($('#element').data('Parsley')).to.be(undefined);
+            expect($('#field1').data('Parsley')).to.be(undefined);
+
+
+            $.unsubscribeAll('parsley:form:validate');
+            done();
+          });
+
+          $('#element').submit();
         });
 
         expect($('#element').data('Parsley')).to.have.key('__class__');
@@ -93,9 +122,6 @@ define(function () {
         expect($('#field1').data('Parsley').__class__).to.be('ParsleyField');
 
         parsleyForm.destroy();
-
-        expect($('#element').data('Parsley')).to.be(undefined);
-        expect($('#field1').data('Parsley')).to.be(undefined);
       });
       afterEach(function () {
         if ($('#element').length)